backend/services/auth/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
pydantic-settings==2.9.1
redis==6.2.0
httpx==0.28.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
pytest==8.3.5
pytest-asyncio==0.25.3
